% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{angle}
\alias{angle}
\title{Angle}
\usage{
angle(x, y)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}
}
\value{
Returns angle between vectors.
}
\description{
Computes the angle between two vectors.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x = c(1,2,3)
y = c(4,5,6)
angle(x,y)
}
\keyword{Helper,}
\keyword{angle,}
\keyword{vectors}
