% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{angle_2d}
\alias{angle_2d}
\title{Angle 2D}
\usage{
angle_2d(x, y)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}
}
\value{
Returns angle between vectors.
}
\description{
Computes the angle between two 2D vectors.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x = c(1,3)
y = c(4,6)
angle_2d(x,y)
}
\keyword{Helper,}
\keyword{angle,}
\keyword{vectors}
