% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_measures.R
\name{compute_jrd_3d_error}
\alias{compute_jrd_3d_error}
\title{Compute JRD 3D Error}
\usage{
compute_jrd_3d_error(jrd_data, destinations)
}
\arguments{
\item{jrd_data}{Data.frame containing the task combo as indices into the destinations list in the description, the result in the value}

\item{destinations}{Data.frame containing the destination name in description and an x,y,z coordinate for each.}
}
\description{
This function extracts the angular error in a 3D
JRD task as defined in EVE.
}
\examples{

destinations <- data.frame(destination=c("Blacksmith workshop", "Gift shop", "Smokehouse", "Cooking shed", "Town hall", "Viking boat", "Wood storage", "Bakery"),
                           x=c(-39,28.5,-38,49.5,12.5,-67,-67,32.5),
                           y=c(6,6.5,5.5,2.5,5.5,4,5.5,4.5),
                           z=c(-23,-14.5,22,48.5,-5.5,31,0,24))
jrd_data <- get_sensors_data_by_device_name(db, device.name = "JRD_3D") \%>\% collect()

jrd_data$error <- compute_jrd_3d_error(jrd_data, destinations)

}
\keyword{JRD,}
\keyword{angular}
\keyword{error}
