% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{deg2rad}
\alias{deg2rad}
\title{Degree To Radian}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{degrees to be converted.}
}
\value{
Radians computed.
}
\description{
Transform degree to radian.
}
\details{
Note: Suggestion by https://stackoverflow.com/questions/32370485/
}
\keyword{Helper,}
\keyword{angle,}
\keyword{unit}
