% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_questions.R
\name{get_participant_answers_ids}
\alias{get_participant_answers_ids}
\title{Get Participant Answer IDs}
\usage{
get_participant_answers_ids(db, session.id = 45)
}
\arguments{
\item{db}{dyplr database handle}

\item{question.name}{name of question as identified in xml}
}
\value{
question and answer of participant
}
\description{
This function extracts a list of a participant's answers ids
from the database. The evaluation is lazy and this data can be
used as input for further remote computations.
}
\examples{
answer_ids <- get_participant_answers_ids(db,session.id=45)


}
\keyword{database,}
\keyword{lazy,}
\keyword{questionnaire}
\keyword{session,}

