% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_times.R
\name{get_participant_duration}
\alias{get_participant_duration}
\title{Get Participant Times}
\usage{
get_participant_duration(db, session.id = 1, scene.id = 0)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session from which to load data}

\item{scene.id}{scene from which to load the data}
}
\value{
Duration of a single scene
}
\description{
This function computes the duration a participant spent in an
experiment scene
}
\examples{
positions <- get_participant_positions_YXZ(session.id = 5)

}
\keyword{EVE,}
\keyword{duration}
\keyword{evaluation,}
