% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_paths.R
\name{get_participant_path_length}
\alias{get_participant_path_length}
\title{Get Participant Path Length}
\usage{
get_participant_path_length(db, session.id = 1, scene.id = 0)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session from which to load data}

\item{scene.id}{scene from which to load the data}
}
\value{
Length of a single path
}
\description{
This function extract the path length for a
given participant's session and scene.
}
\examples{
db <- setup_evertools()
path <- get_participant_path_length(db, session.id = 1, scene.id = 0)

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}
