% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_positions.R
\name{get_participant_positions_XZ}
\alias{get_participant_positions_XZ}
\title{Get Participant Positions (XYZ)}
\usage{
get_participant_positions_XZ(db, session.id = 1,
  scene.name = "Tolman_01")
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session from which to load data}

\item{scene.id}{scene from which to load the data}
}
\description{
This function extracts a participants position (x,z) and time from
the database. The evaluation is lazy and this data can be
used as input for further remote computations.
Note that the Y axis corresponds to the Up Vector in unity data.
This function corresponds to position in a projection to a plane 
with the Up vector as a normal.
}
\examples{
positions <- get_participant_positions_YXZ(db, session.id = 1, scene.name = "Tolman_01")

}
\keyword{database,}
\keyword{lazy,}
\keyword{position}
\keyword{session,}
