% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_times.R
\name{get_participant_times}
\alias{get_participant_times}
\title{Get Participant Times}
\usage{
get_participant_times(db, session.id = 1, scene.id = 0)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session from which to load data}

\item{scene.id}{scene from which to load the data}
}
\value{
time in
}
\description{
This function extracts a participant's time from
the database. The evaluation is lazy and this data can be
used as input for further remote computations.
}
\examples{
positions <- get_participant_positions_YXZ(session.id = 5)

}
\keyword{database,}
\keyword{lazy,}
\keyword{position}
\keyword{session,}
