% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_questions.R
\name{get_participants_bool_answers}
\alias{get_participants_bool_answers}
\title{Get Participants Boolean Answer}
\usage{
get_participants_bool_answers(db, session.ids = c(45),
  question.name = "socDem1")
}
\arguments{
\item{db}{dyplr database handle}

\item{question.name}{name of question as identified in xml}

\item{session.id}{session from which to load data}
}
\value{
question and answer of participant
}
\description{
This function extracts participants' answers from
the database.
}
\examples{
question <- get_participants_bool_answers(db,session.id = 45,question.name="socDem1")

}
\keyword{database,}
\keyword{questionnaire}
\keyword{session,}

