% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_times.R
\name{get_participants_duration}
\alias{get_participants_duration}
\title{Get Participants Duration}
\usage{
get_participants_duration(db, session.ids = c(1), scene.id = 0)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}

\item{scene.id}{scene from which to load the data}
}
\value{
A table with scene ids, session ids and durations.
}
\description{
This function extract the path length for a set of
participants based on session and one scene.
}
\examples{
db <- setup_evertools()
paths <- get_participants_path_length(db, session.ids = c(1:8), scene.ids = 0)

}
\keyword{EVE,}
\keyword{duration}
\keyword{evaluation,}

