% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_times.R
\name{get_participants_duration_all}
\alias{get_participants_duration_all}
\title{Get Participants Duration All}
\usage{
get_participants_duration_all(db, session.ids = c(1), scene.ids = c(0))
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}

\item{scene.ids}{scenes from which to load the data}
}
\value{
A table with session ids and durations.
}
\description{
This function extract the path length for a set of
participants based on sessions and scenes.
}
\examples{
db <- setup_evertools()
paths <- get_participants_path_length_all(db, session.ids = c(1:8), scene.ids = c(0:3))

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}
