% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_paths.R
\name{get_participants_path_length}
\alias{get_participants_path_length}
\title{Get Participants Path Length}
\usage{
get_participants_path_length(db, session.ids = c(1),
  scene.name = "Tolman_01")
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}

\item{scene.id}{scene from which to load the data}
}
\value{
A table with scene ids, session ids and path lengths.
}
\description{
This function extract the path length for a set of
participants based on session and one scene.
}
\examples{
db <- setup_evertools()
paths <- get_participants_path_length(db, session.ids = c(1:8), scene.ids = 0)

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}
