% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_helper.R
\name{get_questions_as_data_frame}
\alias{get_questions_as_data_frame}
\title{Get Questions as Data Frame}
\usage{
get_questions_as_data_frame(db, session.ids, participant.ids = NULL)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}

\item{participant.ids}{optional, identifier for participants}
}
\value{
all answers of all participants
}
\description{
This function extracts all participants' answer from
the database.
}
\examples{
questions.df <- get_questions_as_data_frame(db,session.ids = session.ids ,participant.ids = participant.ids)

}
\keyword{database,}
\keyword{questionnaire}
\keyword{session,}
