% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scene_helper.R
\name{get_scene_id_by_name}
\alias{get_scene_id_by_name}
\title{Get Scene ID By Name}
\usage{
get_scene_id_by_name(db, scene.id = 0)
}
\arguments{
\item{db}{dyplr database handle}

\item{scene.id}{scene from which to load the data}
}
\description{
This function extracts a a scene id by the scene
name.
}
\examples{
scene_name <- get_scene_id_by_name(db, scene.id = 5)

}
\keyword{database,}
\keyword{lazy,}
\keyword{orientation}
\keyword{session,}
