% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_sensor_data.R
\name{get_sensor_data_3d}
\alias{get_sensor_data_3d}
\title{Get Sensor Data 3D}
\usage{
get_sensor_data_3d(db, session.id = 0, sensor.id = 0, start.time = 0,
  end.time = 0)
}
\arguments{
\item{db}{dbyplr database handle}

\item{session.id}{session to be analysed}

\item{sensor.id}{sensor to be analysed}

\item{start.time}{beginning of time segment}

\item{end.time}{end of time segment, ignored if 0.}
}
\description{
This function extracts data for a given
3d sensor for a given time spawn.
}
\examples{
data <- get_sensor_data_3d(db,  session.id = 0, sensor.id = 0, start.time=0, end.time=0)

}
\keyword{database,}
\keyword{lazy,}
\keyword{scene}
\keyword{session,}
