% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_helper.R
\name{get_sensor_data_by_scene}
\alias{get_sensor_data_by_scene}
\title{Get Sensor Data 3D By Scene}
\usage{
get_sensor_data_by_scene(db, session.id = 0, sensor.name = "position",
  scene.name = "The_Viking_Village")
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session to be analysed}

\item{sensor.name}{sensor to be loaded}

\item{scene.name}{scene to be loaded}
}
\description{
This function extracts data for a given
3d sensor for a given scene.
}
\examples{
data <- get_sensor_data_3d_by_scene(db,  session.id = 0, sensor.name = "position", scene.name="The_Viking_Village")

}
\keyword{database,}
\keyword{lazy,}
\keyword{scene}
\keyword{session,}
