% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_helper.R
\name{get_sensor_ids_by_device_name}
\alias{get_sensor_ids_by_device_name}
\title{Get Sensor Ids By Device Name}
\usage{
get_sensor_ids_by_device_name(db, device.name = "JRD")
}
\arguments{
\item{db}{dyplr database handle}

\item{device.name}{device to be loaded}
}
\value{
ids of sensors connected to a device.
}
\description{
This function extracts sensor ids attached
to a particular device.
}
\examples{
data <- get_sensor_id_by_device_name(db, device.name = "JRD")

}
\keyword{database,}
\keyword{device,}
\keyword{lazy,}
\keyword{sensors}
\keyword{session,}
