% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_helper.R
\name{get_sensor_name_by_id}
\alias{get_sensor_name_by_id}
\title{Get Scene Name By ID}
\usage{
get_sensor_name_by_id(db, scene.name = "0")
}
\arguments{
\item{db}{dyplr database handle}

\item{scene.name}{scene to be identified}
}
\description{
This function extracts a a scene name by the scene
id
}
\examples{
scene_name <- get_sensor_name_by_id(db, scene.name = "5")

}
\keyword{database,}
\keyword{lazy,}
\keyword{scene}
\keyword{session,}
