% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_helper.R
\name{get_session_parameters}
\alias{get_session_parameters}
\title{Get Session Parameters}
\usage{
get_session_parameters(db, session.ids = c(45))
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}
}
\value{
table of session parameters
}
\description{
This function extracts the session parameters for all
participants from the database.
}
\examples{
parameters <- get_session_parameters(db, session.ids = c(45))

}
\keyword{database,}
\keyword{parameters}
\keyword{session,}
