% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_helper.R
\name{get_session_scene_time_information}
\alias{get_session_scene_time_information}
\title{Get Session Scene Time Information}
\usage{
get_session_scene_time_information(db, session.id = "0",
  scene.name = "Tolman_01", string.form = F)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.id}{session which to analyse}

\item{scene.name}{scene which to analyse}
}
\description{
This function extracts a a scene id by the scene
name.
}
\examples{
scene_name <- get_session_scene_time_information(db, session.id = "5", scene.name = "Test")
 
}
\keyword{database,}
\keyword{lazy,}
\keyword{orientation}
\keyword{session,}
