% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{ggplotQQ}
\alias{ggplotQQ}
\title{ggplot2 QQ}
\usage{
ggplotQQ(samples)
}
\arguments{
\item{samples}{Data points to be qq plotted.}
}
\value{
A table with session ids and durations.
}
\description{
This function provides a normal Q-Q plot with ggplot2.
}
\details{
The code is originially from here:
http://stackoverflow.com/a/4357932/6260885
}
\examples{
ggplot2_QQ(lm)
paths <- get_participants_path_length_all(db, session.ids = c(1:8), scene.ids = c(0:3))

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}

