% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{\%++\%}
\alias{\%++\%}
\title{Concatenate Operator}
\usage{
s1 \%++\% s2
}
\arguments{
\item{s1}{First string, no trailing spaces.}

\item{s2}{Second string. no leading spaces.}
}
\value{
Concatenated strings with a separating space
}
\description{
This function wraps concatentate
}
\examples{
> s1 <- "a"
> s2 <- "b"
> s1 \%+\% s2
[1] "a b"
}
\keyword{Helper,}
\keyword{string}
