% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{histogram_density_plot_variable}
\alias{histogram_density_plot_variable}
\title{Box Plot Variable}
\usage{
histogram_density_plot_variable(data = NULL, bin.width = 50,
  alpha = 0.2, x.var = NULL, x.name = "", grouping = NULL,
  legend.name = "", legend.labels = NULL, condition.name = "")
}
\arguments{
\item{data}{Data points to be qq plotted.}

\item{bin.width}{Used to compute density.}

\item{alpha}{How transparent the curve is.}

\item{x.var}{Variable to be histogrammed.}

\item{x.name}{Variable name to be shown.}

\item{grouping}{How to group distributions.}

\item{legend.name}{Title of the legend.}

\item{condition.name}{Title of the graph addition.}

\item{legend.label}{New labels for the legend (optional).}
}
\value{
A table with session ids and durations.
}
\description{
This function provides a box plot of a variable.
}
\details{
It can be separated by scene and a controling factor
}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}
