% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{is.Date}
\alias{is.Date}
\title{Is Date}
\usage{
is.Date(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
whether x is Date
}
\description{
Tests whether an object is Date
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x <- "blub"
is.X.Date <- is.Date(x)#returns F
}
\keyword{Helper,}
\keyword{date,}
\keyword{type-testing}
