% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{is.POSIXct}
\alias{is.POSIXct}
\title{Is POSIXct}
\usage{
is.POSIXct(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
whether x is POSIXct
}
\description{
Tests whether an object is POSIXct.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x <- "blub"
is.X.POSIXct <- is.POSIXct(x)#returns F
}
\keyword{Helper,}
\keyword{date,}
\keyword{type-testing}
