% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{is.POSIXlt}
\alias{is.POSIXlt}
\title{Is POSIXlt}
\usage{
is.POSIXlt(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
whether x is POSIXlt
}
\description{
Tests whether an object is POSIXlt.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x <- "blub"
is.X.POSIXlt <- is.POSIXlt(x)#returns F
}
\keyword{Helper,}
\keyword{date,}
\keyword{type-testing}
