% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{is.POSIXt}
\alias{is.POSIXt}
\title{Is POSIXt}
\usage{
is.POSIXt(x)
}
\arguments{
\item{x}{Object to be tested}
}
\value{
whether x is POSIXt
}
\description{
Tests whether an object is POSIXt.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
x <- "blub"
is.X.POSIXt <- is.POSIXt(x)#returns F
}
\keyword{Helper,}
\keyword{date,}
\keyword{type-testing}
