% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{normal_plots}
\alias{normal_plots}
\title{Normal Plots}
\usage{
normal_plots(samples, variable.name = "", scene.name = "")
}
\arguments{
\item{samples}{Data points to be tested for normality.}

\item{variable.name}{Name of the variable which is ploted}

\item{scene.name}{Name of the scene that is ploted}
}
\value{
ggplot2 graph
}
\description{
This function provides a normal Q-Q  and P-P plot with 
ggplot2 with labeling
}
\examples{
paths <- get_participants_path_length_all(db,session.ids = session.ids, scene.ids = scene.ids)
normal_plots(paths$path_length[paths$scene_id==17],"Path Length","Training")

}
\keyword{EVE,}
\keyword{PP-plot,}
\keyword{QQ-plot}
\keyword{evaluation,}
\keyword{normal}
\keyword{normality,}
\keyword{plots,}

