% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{plot_all_participants_by_map}
\alias{plot_all_participants_by_map}
\title{Plot All Participants By Map}
\usage{
plot_all_participants_by_map(db, session.ids = c(1),
  scene.name = "The_Viking_Village", alpha = 1, dataPath = "",
  width = 2048, height = 2048)
}
\arguments{
\item{db}{dbyplr database handle}

\item{session.ids}{Sessions to be loaded}

\item{scene.name}{Name of the scene that is ploted}
}
\value{
ggplot2 graph of paths in scene
}
\description{
This function plots all participants path for a single
scene.
}
\examples{
plot_all_participants_by_map(db, session.ids, "The_Viking_Village", alpha = 0.5, dataPath = unityDataPath)

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{mal,}
\keyword{paths,}
\keyword{top-down}
\keyword{visualisation}
