% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{point2image}
\alias{point2image}
\title{Point To Image}
\usage{
point2image(x, projection, worldToCamera)
}
\arguments{
\item{x}{Vector in 3D.}

\item{projection}{Matrix describing the camera projection.}

\item{worldToCamera}{Matrix describing the camera position.}
}
\value{
x in NPC space.
}
\description{
Transforms a 3 point based on a projection and
world to camera matrix into the NPC space.
}
\examples{
x <- c(1,2,3)
projection <- matrix(c(1,0,0,0,0,1,0,0,0,0,1,-1,0,0,0,1),nrow=4,ncol = 4, byrow = TRUE)
worldToCamera <- matrix(c(-1,0,0,-25,0,0,-1,0,0,1,0,-30,0,0,0,1),nrow=4,ncol = 4, byrow = TRUE)
npc_coordinates <- point2image(x)
}
\keyword{3D}
\keyword{Helper,}
\keyword{projection,}
\keyword{transformation,}
\keyword{vector}
