% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{rad2deg}
\alias{rad2deg}
\title{Radian To Degree}
\usage{
rad2deg(rad)
}
\arguments{
\item{rad}{radian to be converted.}
}
\value{
Degrees computed.
}
\description{
Transform radian to degree.
}
\details{
Note: Suggestion by https://stackoverflow.com/questions/32370485/
}
\keyword{Helper,}
\keyword{angle,}
\keyword{unit}
