% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_evertools.R
\name{setup_evertools}
\alias{setup_evertools}
\title{Setup EVE R Tools}
\usage{
setup_evertools(db = "mysql", dbname = "virtual_study",
  host = "localhost", user = "user", password = "password", port = 3306)
}
\arguments{
\item{db}{Type of database. Defaults to "mysql".}
}
\value{
Dplyr handle for the database12
}
\description{
This function loads basic R libraries for the 
use with EVE and sets up the connection with
the database
}
\examples{
setup_evertools()

}
\keyword{EVE,}
\keyword{database}
