% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{transform_3d_positions_to_image}
\alias{transform_3d_positions_to_image}
\title{Transform 3D Positions To Image}
\usage{
transform_3d_positions_to_image(position_data, projection, worldToCamera)
}
\arguments{
\item{projection}{Matrix describing the camera projection.}

\item{worldToCamera}{Matrix describing the camera position.}

\item{x}{Path data.}
}
\value{
Returns transformed path.
}
\description{
Takes an data.frame of positions (coded with x,y,z)
and returns their position in NPC space.
}
\details{
Note: Suggestion by https://stackoverflow.com/a/26413765
}
\examples{
projection <- matrix(c(1,0,0,0,0,1,0,0,0,0,1,-1,0,0,0,1),nrow=4,ncol = 4, byrow = TRUE)
worldToCamera <- matrix(c(-1,0,0,-25,0,0,-1,0,0,1,0,-30,0,0,0,1),nrow=4,ncol = 4, byrow = TRUE)
position_data <- get_participants_positions_XYZ(db = db, session.ids = session.ids, scene.name = scene.name )
position_data <- transform_3d_positions_to_image(position_data,projection,worldToCamera)
}
\keyword{Helper,}
\keyword{NPC}
\keyword{path}
\keyword{projection}
\keyword{transformation,}
