% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_helper.R
\name{year2age}
\alias{year2age}
\title{Year to Age}
\usage{
year2age(df, base_year)
}
\arguments{
\item{df}{Data frame with age in "age".}

\item{base_year}{Year the experiment took place}
}
\value{
Data frame with years converted to age
}
\description{
This function converts years to ages. Assumes
that age is in the column "age".
}
\details{
Note: Assumes that humans do not grow older than 150.
}
\examples{
df <- year2age(df,2016)
}
\keyword{Helper,}
\keyword{age,}
\keyword{conversion}
\keyword{year,}

