% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe_helper.R
\name{get_questions_as_data_frame}
\alias{get_questions_as_data_frame}
\title{Get Questions as Data Frame}
\usage{
get_questions_as_data_frame(db, session.ids, participant.ids = NULL,
  utf8 = T)
}
\arguments{
\item{db}{dyplr database handle}

\item{session.ids}{sessions from which to load data}

\item{participant.ids}{optional, identifier for participants}

\item{utf8}{optional allows to turn of encoding in utf8}
}
\value{
all answers of all participants
}
\description{
This function extracts all participants' answers from
the database. Note that this exports the answer indices
rather than the output values. So the value 0 describes
the first output.
}
\examples{
questions.df <- get_questions_as_data_frame(db,session.ids = session.ids ,participant.ids = participant.ids)
}
\keyword{database,}
\keyword{export}
\keyword{questionnaire,}
