% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_helper.R
\name{get_sensor_id_by_name}
\alias{get_sensor_id_by_name}
\title{Get Scene ID By Name}
\usage{
get_sensor_id_by_name(db, sensor.name = "0")
}
\arguments{
\item{db}{dyplr database handle}

\item{scene.name}{scene to be identified}
}
\description{
This function extracts a a scene name by the scene
id
}
\examples{
sensor_id <- get_sensor_id_by_name(db, scene.name = "5")

}
\keyword{database,}
\keyword{lazy,}
\keyword{scene}
\keyword{session,}
