% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helper.R
\name{histogram_density_plot_variable}
\alias{histogram_density_plot_variable}
\title{Box Plot Variable}
\usage{
histogram_density_plot_variable(data = NULL, bin.width = 50, alpha = 0.2,
  x.var = NULL, x.name = "", grouping = NULL, legend.name = "",
  legend.labels = NULL, condition.name = "")
}
\arguments{
\item{samples}{Data points to be qq plotted.}

\item{variable.name}{Name of the variable which is ploted}

\item{scene.name}{Name of the scene that is ploted}
}
\value{
A table with session ids and durations.
}
\description{
This function provides a box plot of a variable.
}
\details{
It can be separated by scene and a controling factor
}
\examples{
#Box plot without variable subgroupings
box_plot_variable(test, 
                  y.var = "path_length", 
                  x.var = "scene_name", 
                  grouping = "scene_id", 
                  y.name = "Path Length", 
                  x.name = "Scene", 
                  group.label = c("Training","City"))
#Box plot with variable subgroupings
box_plot_variable(test,
                  y.var = "path_length",
                  x.var = "scene_name",
                  grouping = "scene_id + sex",
                  coloring = "sex",
                  y.name = "Path Length",
                  x.name = "Scene",
                  group.label = c("Training","City"),
                  legend.name = "Sex",
                  legend.label = c("TRUE"="Female", "FALSE"="Male"))

}
\keyword{EVE,}
\keyword{evaluation,}
\keyword{path}
